Added past and future states to variables (closes #15) Fixed error handling (closes #23) Changed commands to get the original entered text (in preparation for #4)

+++ src/ch/njol/skript/data/DefaultChangers.java
@@ -190,7 +190,7 @@ public class DefaultChangers {
 				case ADD:
 					final Timespan ts = (Timespan) delta.getSingle(e);
 					for (final World w : worlds.getArray(e)) {
-						w.setTime((long) (w.getTime() + mod * ts.getTicks()));
+						w.setTime(w.getTime() + mod * ts.getTicks());
 					}
 				break;
 			}
